Refactor debug UI into separate module
Separated Debug UI menu code out from egui_integration
