+++ src/engine/ui.rs
@@ -1,1 +1,2 @@
 pub mod egui_integration;
+mod debug_ui;
